g_object_get: Fall back to g_object_get_valist if something is wrong
g_object_get() is core GLib functionality and is called by non-GTK
programs, including non-GUIs. To minimize the impact on non-GTK programs
and the chance of causing a crash, follow the principle of "first, do
no harm" and only enter gtk3-nocsd's custom code path when everything
is as it is normally expected to be:

* a non-NULL object pointer
* an object pointer that is, in fact, a valid GObject
* and a compatible version of GTK loaded in memory

In all other cases, just pass the arguments directly on to
`g_object_valist()`.

As well as being safer, this has the side benefit of restoring the
warning that `g_object_get()` would normally try to produce when called
on an invalid object (null pointer or use-after-free), which is a useful
debugging aid.

Signed-off-by: Simon McVittie <smcv@collabora.com>